leave last[XY] vars as ints
git-svn-id: https://aeon.stsci.edu/ssb/svn/pyraf/trunk@812 05fa7843-0683-0410-9e38-dfee850dc5be

+++ lib/gkitkbase.py
@@ -685,8 +685,8 @@ class GkiInteractiveTkBase(gki.GkiKernel, wutil.FocusEntity):
         gwidget = self.gwidget
         if gwidget:
             if gwidget.lastX is None:
-                gwidget.lastX = gwidget.winfo_width()/2
-                gwidget.lastY = gwidget.winfo_height()/2
+                gwidget.lastX = int(gwidget.winfo_width()/2.)
+                gwidget.lastY = int(gwidget.winfo_height()/2.)
             wutil.moveCursorTo(gwidget.winfo_id(),gwidget.lastX,gwidget.lastY)
             gwidget.focus_force()
